requirements.txt
jax>=0.4.8
jaxlib>=0.4.7
chex>=0.1.7
evosax>=0.1.4
dm-haiku>=0.0.6
optax>=0.1.2
xarray>=2022.3.0
PyYAML>=6.0
matplotlib>=3.7.0
scikit-learn>=1.1.1
tqdm>=4.64.0
pytest>=7.1.2
netCDF4>=1.6.0
scikit-optimize==0.9.0
dm_pix>=0.4.0
cma==3.3.0
